Migrate data-related code from ServerListPresenter to ServerListDataManager.
