+++ app/src/main/java/io/github/trojan_gfw/igniter/servers/data/ServerListDataManager.java
@@ -1,9 +1,15 @@
 package io.github.trojan_gfw.igniter.servers.data;
 
+import android.content.Context;
+import android.net.Uri;
 import android.text.TextUtils;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -30,6 +36,7 @@ import javax.net.ssl.HttpsURLConnection;
 import io.github.trojan_gfw.igniter.TrojanConfig;
 import io.github.trojan_gfw.igniter.TrojanHelper;
 import io.github.trojan_gfw.igniter.TrojanURLHelper;
+import io.github.trojan_gfw.igniter.common.constants.ConfigFileConstants;
 import io.github.trojan_gfw.igniter.common.utils.DecodeUtils;
 
 public class ServerListDataManager implements ServerListDataSource {
@@ -214,5 +221,100 @@ public class ServerListDataManager implements ServerListDataSource {
             e.printStackTrace();
         }
         return null;
+    }
+
+    @Nullable
+    private String readFileContentFromUri(Context context, Uri fileUri) {
+        try (InputStream is = context.getContentResolver().openInputStream(fileUri)) {
+            return readStringFromStream(is);
+        } catch (IOException e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
+
+    @Override
+    public List<TrojanConfig> importServersFromFile(Context context, Uri fileUri) {
+        List<TrojanConfig> trojanConfigs = parseTrojanConfigsFromFileContent(readFileContentFromUri(context, fileUri));
+        List<TrojanConfig> currentConfigs = loadServerConfigList();
+        currentConfigs.addAll(trojanConfigs);
+        // remove repeated configurations
+        Set<String> newTrojanConfigRemoteAddrSet = new HashSet<>();
+        for (TrojanConfig config : trojanConfigs) {
+            newTrojanConfigRemoteAddrSet.add(config.getRemoteAddr());
+        }
+        for (int i = currentConfigs.size() - 1; i >= 0; i--) {
+            if (newTrojanConfigRemoteAddrSet.contains(currentConfigs.get(i).getRemoteAddr())) {
+                currentConfigs.remove(i);
+            }
+        }
+        currentConfigs.addAll(trojanConfigs);
+        replaceServerConfigs(currentConfigs);
+        return currentConfigs;
+    }
+
+    @NonNull
+    private List<TrojanConfig> parseTrojanConfigsFromFileContent(String fileContent) {
+        try {
+            JSONObject jsonObject = new JSONObject(fileContent);
+            JSONArray configs = jsonObject.optJSONArray(ConfigFileConstants.CONFIGS);
+            if (configs == null) {
+                return Collections.emptyList();
+            }
+            final int len = configs.length();
+            List<TrojanConfig> list = new ArrayList<>(len);
+            for (int i = 0; i < len; i++) {
+                JSONObject config = configs.getJSONObject(i);
+                String remoteAddr = config.optString(ConfigFileConstants.SERVER, null);
+                if (remoteAddr == null) {
+                    continue;
+                }
+                TrojanConfig tmp = new TrojanConfig();
+                tmp.setRemoteServerRemark(config.optString(ConfigFileConstants.REMARKS, ConfigFileConstants.NO_REMARKS));
+                tmp.setRemoteAddr(remoteAddr);
+                tmp.setRemotePort(config.optInt(ConfigFileConstants.SERVER_PORT));
+                tmp.setPassword(config.optString(ConfigFileConstants.PASSWORD));
+                tmp.setVerifyCert(config.optBoolean(ConfigFileConstants.VERIFY));
+                list.add(tmp);
+            }
+            return list;
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+        return Collections.emptyList();
+    }
+
+    @Override
+    public boolean exportServers(String exportPath) {
+        return TrojanHelper.writeStringToFile(getExportContent(), exportPath);
+    }
+
+    private String getExportContent() {
+        List<TrojanConfig> trojanConfigs = loadServerConfigList();
+        JSONArray array = new JSONArray();
+        int index = 0;
+        for (TrojanConfig trojanConfig : trojanConfigs) {
+            JSONObject jsonObject = new JSONObject();
+            try {
+                jsonObject.put(ConfigFileConstants.REMARKS, trojanConfig.getRemoteServerRemark());
+                jsonObject.put(ConfigFileConstants.SERVER, trojanConfig.getRemoteAddr());
+                jsonObject.put(ConfigFileConstants.SERVER_PORT, trojanConfig.getRemotePort());
+                jsonObject.put(ConfigFileConstants.PASSWORD, trojanConfig.getPassword());
+                jsonObject.put(ConfigFileConstants.VERIFY, trojanConfig.getVerifyCert());
+                // for future
+                // jsonObject.put("enable_ipv6", trojanConfig.getEnableIpv6());
+                // jsonObject.put("enable_clash", trojanConfig.getEnableClash());
+                array.put(index++, jsonObject);
+            } catch (JSONException e) {
+                e.printStackTrace();
+            }
+        }
+        JSONObject jsonObject = new JSONObject();
+        try {
+            jsonObject.put(ConfigFileConstants.CONFIGS, array);
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+        return jsonObject.toString();
     }
 }
