+++ app/src/main/java/io/github/trojan_gfw/igniter/servers/data/ServerListDataSource.java
@@ -1,5 +1,8 @@
 package io.github.trojan_gfw.igniter.servers.data;
 
+import android.content.Context;
+import android.net.Uri;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.WorkerThread;
 
@@ -21,6 +24,19 @@ public interface ServerListDataSource {
     void replaceServerConfigs(List<TrojanConfig> list);
     @WorkerThread
     void requestSubscribeServerConfigs(String url, @NonNull Callback callback);
+
+    /**
+     * Parse trojan configs from {@param fileUri}. Combine with current trojan config list and return
+     * the complete list.
+     *
+     * @param context Context
+     * @param fileUri File Uri
+     * @return List of all trojan configs.
+     */
+    @WorkerThread
+    List<TrojanConfig> importServersFromFile(Context context, Uri fileUri);
+    @WorkerThread
+    boolean exportServers(String exportPath);
 
     interface Callback {
         void onSuccess();
